microbian: implement safe multi-processor operation
This commit introduces basic multiprocessor support into micro:bian. The
idea is very simple: all kernel code is protected by a shared lock,
which is acquired whenever kernel code is entered via an interrupt and
released before returning to user code. This is implemented for the Pico
using an RP2040 SIO spinlock.

Only core 0 is permitted to receive interrupts; as such, processes which
connect to interrupts are constrained to core 0. This is achieved in the
scheduler, by ignoring processes with priority `P_HANDLER` in auxiliary
cores.

The entry point of the OS itself (called from `startup.c`) has changed.
Instead of `__start`, there are now two functions: `__init` and
`__start_core`. The former must be called first, and only once it
returns can all cores - including core 0 - call the latter to begin
their scheduling loops.