Implement debug pins to visualize active tasks
This change introduces "debug pins" into various parts of the codebase:
GPIO pins which are pulled high whilst some condition is met, such as
contention on the kernel lock. This mechanism seves as a debugging tool:
using a logic analyzer, one can see the actions of each CPU core with
minimal impact on the functionality of the system (since GPIO writes are
generally fast).